==> Data_Extraction_and_Merging/1_Other_Dataset_Extraction.ipynb <==
import trafilatura
import csv
import requests
from datetime import datetime
from urllib.parse import urlparse
from openpyxl import Workbook

# Define the URL of the article
url = "https://www.telegraph.co.uk/columnists/2023/07/13/asylum-seeker-pretended-to-be-child-robert-jenrick/"
# Download the content of the article using requests
downloaded = requests.get(url).text

# Check if the request was successful
#if response.status_code == 200:
#downloaded = response.text

# Extract the content of the article using trafilatura
content = trafilatura.extract(downloaded)

# Extract the publication date from the URL (current date in this case)
publication_date = datetime.now().strftime("%Y-%m-%d")

# Parse the URL to extract the domain name
website_name = urlparse(url).netloc

# Create a new Excel workbook
workbook = Workbook()
sheet = workbook.active

# Write headers
sheet["A1"] = "Publication Date"
sheet["B1"] = "Website Name"
sheet["C1"] = "Content"

# Write data
sheet.append([publication_date, website_name, content])

# Save the workbook to an Excel file
workbook.save(r"C:\Users\pc\Desktop\Extracts\2490_data.xlsx")

#else:
    #print(f"Failed to retrieve the URL. Status code: {response.status_code}")
