==> Data_Extraction_and_Merging/3_Guardian_Data_Manipulation.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("/Users/henryasiamah/Desktop/work/newone_guardian_news.csv")
df.head()

df.count()

# Text Cleaning

# finding the total sum of duplicates
df.duplicated().sum()

#viewing the total sum of duplicates
df[df.duplicated()]

df.shape

# # Define keywords to filter titles
# keywords = ["Refugees", "Migration", "Immigration", "Refugee", "Asylum Seeker", "Asylum Seekers", "Migrants", "Migrant", "Immigrants", "Immigrant"]

# # Create a regex pattern that matches any of the keywords
# pattern = '|'.join(keywords)

# # Filter rows where 'webTitle' contains any of the keywords (case-insensitive)
# filtered_df = df[df['webTitle'].str.contains(pattern, case=False, na=False)]

# # # Display or save the filtered DataFrame
# # filtered_df.to_csv("filtered_file.csv", index=False)  # Save to a new CSV if needed

# filtered_df.head()

# df = df.loc[df['publication_date_fixed'] >= '2021-01-01']

# df.head()

#filtered_df.tail()

#filtered_df.shape

df["News Paper Source"] = "The Guardian"

df.head()

# Extract data post corona
df = df.loc[df['publication_date_fixed'] >= '2021-01-01']

df.head()

df.tail()

df.to_csv("/Users/henryasiamah/Desktop/work/2021_guardian_news.csv", index = False)

df = pd.read_csv("/Users/henryasiamah/Desktop/work/2021_guardian_news.csv")
df.head()

df.tail()

df.shape

df.tail()

# Define keywords to filter titles
keywords = ["Refugees", "Migration", "Immigration", "Refugee", "Asylum Seeker", "Asylum Seekers", "Migrants", "Migrant", "Immigrants", "Immigrant"]

# Create a regex pattern that matches any of the keywords
pattern = '|'.join(keywords)

# Filter rows where 'webTitle' contains any of the keywords (case-insensitive)
filtered_df = df[df['webTitle'].str.contains(pattern, case=False, na=False)]

# Display or save the filtered DataFrame
#filtered_df.to_csv("filtered_file.csv", index=False)  # Save to a new CSV if needed

filtered_df.head()

filtered_df.shape

filtered_df.tail()

filtered_df.to_csv("/Users/henryasiamah/Desktop/work/filtered_guardian_news.csv", index = False)

filtered = pd.read_csv("/Users/henryasiamah/Desktop/work/filtered_guardian_news.csv")
filtered.head()

filtered.tail()

filtered.shape
