==> Data_Extraction_and_Merging/4_Other_dataset_ Manipulation.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("/Users/henryasiamah/Desktop/work/other_dataset.csv")
df.head()

df.shape

df['Website Name'].value_counts()

df.info()

"""A fucntion to convert the Website Name column which
    contains the News Paper Source but in urls to the 
    News Paper Source name. eg. www.telegraph.co.uk becomes  The Telegrah UK
"""

def source(value):
    if value == "www.telegraph.co.uk":
        return "The Telegraph"
    elif value == "www.dutchnews.nl":
        return "Dutch News"
    elif value == "www.thelocal.de":
        return "The Local (Germany)"
    elif value == "www.thelocal.it":
        return "The Local (Italy)"
    elif value == "www.thelocal.fr":
        return "The Local (France)"
    elif value == "www.ansa.it":
        return "ANSA"
    elif value == "www.thelocal.es":
        return "The Local (Spain)"
    elif value == "www.infomigrants.net":
        return "InfoMigrants"
    elif value == "www.euronews.com":
        return "Euronews"
    elif value == "www.dw.com":
        return "Deutsche Welle (DW)"
    elif value == "www.bbc.com":
        return "BBC"
    elif value == "www.france24.com":
        return "France 24"
    elif value == "nltimes.nl":
        return "NL Times"
    elif value == "www.lemonde.fr":
        return "Le Monde"
    elif value == "www.thenewhumanitarian.org":
        return "The New Humanitarian"
    elif value == "www.politico.eu":
        return "Politico Europe"
    elif value == "immigrationlab.org":
        return "Immigration Policy Lab"
    elif value == "www.theguardian.com":
        return "The Guardian"
    elif value == "freemovement.org.uk":
        return "Free Movement"
    elif value == "www.aljazeera.com":
        return "Euronews"
    elif value == "apnews.com":
        return "Euronews"
    elif value == "home-affairs.ec.europa.eu":
        return "European Commission Home Affairs"
    elif value == "www.irishtimes.com":
        return "The Irish Times"
    elif value == "www.chathamhouse.org":
        return "Chatham House"
    elif value == "theconversation.com":
        return "Euronews"
    elif value == "www.auswaertiges-amt.de":
        return "German Federal Foreign Office"
    elif value == "eumigrationlawblog.eu":
        return "EU Migration Law Blog"
    elif value == "www.iemed.org":
        return "IEMed"
    elif value == "www.voanews.com":
        return "Immigration Policy Lab"
    elif value == "www.hrw.org":
        return "Immigration Policy Lab"
    elif value == "www.unicef.org":
        return "UNICEF"
    elif value == "www.migrationpolicy.org":
        return "Immigration Policy Lab"
    elif value == "syacd.org":
        return "Immigration Policy Lab"
    elif value == "edition.cnn.com":
        return "Immigration Policy Lab"
    elif value == "www.context.news":
        return "Context News-TRF"
    elif value == "worldmigrationreport.iom.int":
        return "IOM"
    elif value == "news.un.org":
        return "United Nations News"
    elif value == "ecre.org":
        return "ECRE"
    else:
        return "Unknown Source"

df["News Paper Source"] = df["Website Name"].apply(source)
df.head()

df.tail()

df["News Paper Source"].value_counts()

"""A fucntion showing the country or location newspaper source is from
"""

def news_source_country(value):
    if value == "The Telegraph":
        return "UK"
    elif value == "Dutch News":
        return "Netherlands"
    elif value == "The Local (Germany)":
        return "Germany"
    elif value == "The Local (Italy)":
        return "Italy"
    elif value == "The Local (France)":
        return "France"
    elif value == "ANSA":
        return "Italy"
    elif value == "The Local (Spain)":
        return "Spain"
    elif value == "Euronews":
        return "France"
    elif value == "InfoMigrants":
        return "Germany"
    elif value == "Deutsche Welle (DW)":
        return "Germany"
    elif value == "BBC":
        return "UK"
    elif value == "France 24":
        return "France"
    elif value == "NL Times":
        return "Netherlands"
    elif value == "Le Monde":
        return "France"
    elif value == "The New Humanitarian":
        return "Switzerland"
    elif value == "Immigration Policy Lab":
        return "Switzerland"
    elif value == "Politico Europe":
        return "Belgium"
    elif value == "The Guardian":
        return "UK"
    elif value == "Free Movement":
        return "UK"
    elif value == "The Irish Times":
        return "Ireland"
    elif value == "Chatham House":
        return "UK"
    elif value == "European Commission Home Affairs":
        return "European Union"
    elif value == "German Federal Foreign Office":
        return "Germany"
    elif value == "IOM":
        return "Switzerland"
    elif value == "EU Migration Law Blog":
        return "European Union"
    elif value == "IEMed":
        return "Spain"
    elif value == "UNICEF":
        return "International"
    elif value == "Context News-TRF":
        return "UK"
    elif value == "ECRE":
        return "Belgium"
    elif value == "United Nations News":
        return "European Union"
    else:
        return "Unknown Country"

df["Country/Organization"] = df["News Paper Source"].apply(news_source_country)
df.head()

df.to_csv("/Users/henryasiamah/Desktop/work/new_other_dataset.csv", index = False)

df = pd.read_csv("/Users/henryasiamah/Desktop/work/new_other_dataset.csv")
df.head()

df["News Paper Source"].value_counts()

df["Country/Organization"].value_counts()

# rename columns

df = df.rename(columns={'Publication Date': 'Date', 'Website Name': 'Headline', 'News Paper Source': 'Source'})

df.head()

df.to_csv("/Users/henryasiamah/Desktop/work/new_other_dataset.csv", index = False)

df = pd.read_csv("/Users/henryasiamah/Desktop/work/new_other_dataset.csv")
df.head()
