==> Data_Extraction_and_Merging/5_Merged_Text_&_Cleaning_Corpus.ipynb <==
from collections import Counter
import pandas as pd
import re

## Merging the other_dataset and the Guardian Dataset.

# Load both files. 
df1 = pd.read_csv("/Users/henryasiamah/Desktop/work/new_other_dataset.csv")
df2 = pd.read_csv("/Users/henryasiamah/Desktop/work/filtered_guardian_news.csv")

# Rename Columns in df2 to match columns in df1.
df2 = df2.rename(columns={'publication_date_fixed': 'Date', 'webTitle': 'Headline', 'body': 'Content', 'News Paper Source': 'Source'})

# Add a new column to df2: Country/Organization and set it to UK
df2["Country/Organization"] = 'UK'

# Select columns in df1 that matches desired columns in df2.
df1 = df1[['Date', 'Headline', 'Content', 'Source', 'Country/Organization']]

#load df2
df2 = df2[['Date', 'Headline', 'Content', 'Source', 'Country/Organization']]

#Concatenate the two dataframes vertically.
final_df = pd.concat([df1, df2])

#Convert the 'Date' column to datetime format.
final_df['Date'] = pd.to_datetime(final_df['Date'], errors='coerce')

#print the original number of rows and columns for df1 and df2
print("Original number of rows and columns for df1:")
print(df1.shape)
print("Original number of rows and columns for df2:")
print(df2.shape)

#print the number of rows and columns for the final_df or merged dataset.
print("Number of rows and columns for the final dataframe")
print(final_df.shape)

#fill missing values for empty string
final_df.fillna('', inplace=True)

final_df.head()

final_df.tail()

final_df.count()

final_df.to_csv("/Users/henryasiamah/Desktop/work/final_df.csv", index = False)

# Text Cleaning

df = final_df

df.duplicated(subset=['Content']).sum()

df.drop_duplicates(subset=['Content'], inplace=True)

# Remove rows with NaN values in 'Cotent' column
df.dropna(subset=['Content'], inplace=True)

# Remove rows with 'NaN' values in 'Content'column
df = df[~df['Content'].isin(['NaN'])]

#print the cleaned dataset
print('The cleaned dataframe ahs {} rows and {} columns'.format(df.shape[0], df.shape[1]))

df.head(10)

df.tail(10)

## Most Common Words

Creating an overview of the most common words in the corpus.

# Concatenate all the strings in the 'Content' column into a single string
all_text = ' '.join(df['Content'].values)

# Use the counter class from the collections module to count the occurences of each word
word_counts = Counter(all_text.split())

# Get the 1000 most common words and their counts
most_common = word_counts.most_common(10000)

# Convert the results to a Pandas DataFrame for easier manipulation and printing
most_common_df = pd.DataFrame(most_common, columns=['Word', 'Count'])

# Print the DataFrame with the most common words and their counts in descending order
print(most_common_df.sort_values('Count', ascending=False))

most_common_df.to_csv("/Users/henryasiamah/Desktop/work/most_common_words.csv", index = False)

#Define a function to clean the text. 
def clean_text(text):
    
    # Remove inline frame element - REMOVE
    text = re.sub(r'', '', text, flags=re.DOTALL)

     # Remove specified text
    patterns = r'\b(p|dataatomidbddcbffdabfcfcfbe|href|blockquote|span|class|element|element-image_caption|element-image_credit|figcaption|liststrong|Photo/Rodrigo|ADVERTISEMENT|element--supporting|Photo/Evgeniy|element-rich-link|copyright|element--showcase|AD-FREE,|meterActive/meterExpired|Photo/Vadim|VideoWhy|/PRNewswire/|itemV1|data-interactive|element-interactive|allowfullscreen="true"|data-atom-type="media"|element-atom|gu-atom|class="twitter-tweet"p|auto-generated|class="timezone"BST/span/time|class="quoted|class="gu-image"|class="/figcaption"|class="timezone"GMT/span/time|class="block"|data-block-contributor=""|class="block-elements"|class="element|href)\b'
    text = re.sub(patterns, '', text)
    
    # Remove 
    text = re.sub(r'', '', text)
    text = re.sub(r'', '', text)
    text = re.sub(r'', '', text)
    text = re.sub(r'/p', '', text)

    # Remove HTML tags
    text = re.sub(r'<.*?>', '', text)
    
    # Remove '@' symbols
    text = re.sub(r'@', '', text)
    
    # Remove links
    text = re.sub(r'http\S+', '', text)
    
    # Remove emojis
    text = text.encode('ascii', 'ignore').decode('ascii')
    
   
    return text 

df['cleaned_corpus'] = df['Content'].apply(clean_text)

df.head()

df.tail()

df.to_csv("/Users/henryasiamah/Desktop/work/cleaned.csv", index=False)
